+++ src/Symfony/Component/Messenger/EventListener/SendFailedMessageForRetryListener.php
@@ -63,7 +63,12 @@ class SendFailedMessageForRetryListener implements EventSubscriberInterface
 
             ++$retryCount;
 
-            $delay = $retryStrategy->getWaitingTime($envelope, $throwable);
+            $delay = null;
+            if ($throwable instanceof RecoverableExceptionInterface && method_exists($throwable, 'getRetryDelay')) {
+                $delay = $throwable->getRetryDelay();
+            }
+
+            $delay ??= $retryStrategy->getWaitingTime($envelope, $throwable);
 
             $this->logger?->warning('Error thrown while handling message {class}. Sending for retry #{retryCount} using {delay} ms delay. Error: "{error}"', $context + ['retryCount' => $retryCount, 'delay' => $delay, 'error' => $throwable->getMessage(), 'exception' => $throwable]);